Add files via upload
This is version .01. It requires me to have manually downloaded and unzipped all the NVD data files I need. I then have to run the script multiple times to extract data from every year. Once that is done once though, that should be most of the manual work. All that is left is to download the "modified" file periodically. With some Excel magic, including vlookup, I am able to update a master list of cves and their access vectors.